pyproject.toml
[project]
name = "chatbotdemo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "openai>=2.20.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.54.0",
]
